+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/torState/TorRelaySettings.kt
@@ -20,7 +20,7 @@
  */
 package com.vitorpamplona.amethyst.model.torState
 
-import com.vitorpamplona.amethyst.ui.tor.TorType
+import com.vitorpamplona.amethyst.commons.tor.TorType
 
 data class TorRelaySettings(
     val torType: TorType = TorType.OFF,
