+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/torState/TorRelayState.kt
@@ -21,9 +21,9 @@
 package com.vitorpamplona.amethyst.model.torState
 
 import androidx.compose.runtime.Stable
+import com.vitorpamplona.amethyst.commons.tor.TorType
 import com.vitorpamplona.amethyst.service.okhttp.DualHttpClientManager
 import com.vitorpamplona.amethyst.ui.tor.TorSettingsFlow
-import com.vitorpamplona.amethyst.ui.tor.TorType
 import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
